Revert "Abclass"
